Split MOCK_AIS into two files (one for mocked ships, one for real exactAIS ships)
